[MOVE/#7] 파일 위치 이동
