+++ app/src/main/res/layout/fragment_do_android.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.main.DoAndroidFragment">
+    tools:context=".presentation.main.doandroid.DoAndroidFragment">
 
     <TextView
         android:id="@+id/tv_do_android"
