+++ app/src/main/res/layout/fragment_my_page.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_height="match_parent"
     android:layout_width="match_parent"
-    tools:context=".presentation.main.MyPageFragment">
+    tools:context=".presentation.main.mypage.MyPageFragment">
 
 
     <ImageView
